requirements.txt
Appium-Python-Client
pytest
pytest-html
allure-pytest